+++ src/main/java/screens/nodehandlers/INodeHandler.java
@@ -1,4 +1,4 @@
-package screens;
+package screens.nodehandlers;
 
 import datastructure.DrawNode;
