+++ src/main/java/screens/nodehandlers/NodeCenter.java
@@ -1,4 +1,4 @@
-package screens;
+package screens.nodehandlers;
 
 import datastructure.DrawNode;
 import javafx.scene.Scene;
@@ -7,6 +7,7 @@ import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
+import screens.scenes.GraphScene;
 
 /**
  * Implementation of the state that handles click events regarding the center query of selected node.
