+++ src/main/java/screens/nodehandlers/NodeInfo.java
@@ -1,7 +1,8 @@
-package screens;
+package screens.nodehandlers;
 
 import datastructure.DrawNode;
 import datastructure.NodeGraph;
+import screens.Window;
 
 /**
  * Implementation of the state that handles click events regarding the information of selected node.
