+++ src/test/java/screens/scenes/GraphSceneTest.java
@@ -1,4 +1,4 @@
-package screens;
+package screens.scenes;
 
 import datastructure.Node;
 import datastructure.NodeGraph;
@@ -15,6 +15,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import screens.FXElementsFactory;
 
 import javax.swing.*;
 import java.util.ArrayList;
@@ -22,7 +23,11 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.anyString;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 /**
  * Tests the GraphScene class.
